refactor(api): split member.service.ts into services/member/ (api-trlq)
Establishes the pattern for epic api-6l1q. Verb-split files under
services/member/ with a sibling services/member.ts barrel. All 23
callers (routes, tRPC router, integration helpers, tests) updated
to the new import path. No behavior changes.

Layout:
- services/member.ts              — barrel re-exports (sibling file)
- services/member/create.ts       — createMember, duplicateMember
- services/member/queries.ts      — listMembers, getMember, listAllMemberMemberships
- services/member/update.ts       — updateMember
- services/member/lifecycle.ts    — archiveMember, restoreMember, deleteMember
- services/member/internal.ts     — shared toMemberResult helper + MemberResult type

Note on barrel placement: TypeScript with moduleResolution "Bundler"
does not resolve ../../services/member.js to services/member/index.ts.
The barrel must live as a sibling file services/member.ts. Subagents
replicating this pattern must follow this structure.
