+++ apps/api/src/__tests__/trpc/integration-helpers.ts
@@ -27,7 +27,7 @@ import { afterAll, afterEach, beforeAll, beforeEach, expect } from "vitest";
 import { createBucket } from "../../services/bucket.service.js";
 import { generateFriendCode, redeemFriendCode } from "../../services/friend-code.service.js";
 import { createFrontingSession } from "../../services/fronting-session.service.js";
-import { createMember } from "../../services/member.service.js";
+import { createMember } from "../../services/member.js";
 import { createStructureEntity } from "../../services/structure-entity-crud.service.js";
 import { createEntityType } from "../../services/structure-entity-type.service.js";
 import { router } from "../../trpc/trpc.js";
