+++ apps/api/src/routes/members/get.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { getMember } from "../../services/member.service.js";
+import { getMember } from "../../services/member.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
