+++ apps/api/src/routes/members/memberships.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../lib/db.js";
 import { requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { listAllMemberMemberships } from "../../services/member.service.js";
+import { listAllMemberMemberships } from "../../services/member.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
