+++ apps/api/src/routes/members/restore.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { restoreMember } from "../../services/member.service.js";
+import { restoreMember } from "../../services/member.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
