+++ apps/api/src/routes/members/update.ts
@@ -7,7 +7,7 @@ import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { parseJsonBody } from "../../lib/parse-json-body.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { updateMember } from "../../services/member.service.js";
+import { updateMember } from "../../services/member.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
