+++ apps/api/src/services/member.ts
@@ -1,0 +1,6 @@
+export { createMember, duplicateMember } from "./member/create.js";
+export { listMembers, getMember, listAllMemberMemberships } from "./member/queries.js";
+export type { MemberMembershipsResult } from "./member/queries.js";
+export { updateMember } from "./member/update.js";
+export { archiveMember, restoreMember, deleteMember } from "./member/lifecycle.js";
+export type { MemberResult } from "./member/internal.js";
