+++ apps/api/src/services/member/internal.ts
@@ -1,0 +1,38 @@
+import { brandId, toUnixMillis, toUnixMillisOrNull } from "@pluralscape/types";
+
+import { encryptedBlobToBase64 } from "../../lib/encrypted-blob.js";
+
+import type { EncryptedBlob, MemberId, SystemId, UnixMillis } from "@pluralscape/types";
+
+export interface MemberResult {
+  readonly id: MemberId;
+  readonly systemId: SystemId;
+  readonly encryptedData: string;
+  readonly version: number;
+  readonly createdAt: UnixMillis;
+  readonly updatedAt: UnixMillis;
+  readonly archived: boolean;
+  readonly archivedAt: UnixMillis | null;
+}
+
+export function toMemberResult(row: {
+  id: string;
+  systemId: string;
+  encryptedData: EncryptedBlob;
+  version: number;
+  createdAt: number;
+  updatedAt: number;
+  archived: boolean;
+  archivedAt: number | null;
+}): MemberResult {
+  return {
+    id: brandId<MemberId>(row.id),
+    systemId: brandId<SystemId>(row.systemId),
+    encryptedData: encryptedBlobToBase64(row.encryptedData),
+    version: row.version,
+    createdAt: toUnixMillis(row.createdAt),
+    updatedAt: toUnixMillis(row.updatedAt),
+    archived: row.archived,
+    archivedAt: toUnixMillisOrNull(row.archivedAt),
+  };
+}
